+++ planetiler-core/src/test/java/com/onthegomap/planetiler/geo/TileCoordTest.java
@@ -39,8 +39,8 @@ class TileCoordTest {
     "32767,0,15,1431655764",
     "32767,32767,15,1431622997"
   })
-  void testTileCoordEncode(int x, int y, int z, int i) {
-    int encoded = TileCoord.ofXYZ(x, y, z).encoded();
+  void testTileCoordEncode(int x, int y, int z, long i) {
+    long encoded = TileCoord.ofXYZ(x, y, z).encoded();
     assertEquals(i, encoded);
     TileCoord decoded = TileCoord.decode(i);
     assertEquals(decoded.x(), x, "x");
@@ -50,9 +50,9 @@ class TileCoordTest {
 
   @Test
   void testTileSortOrderRespectZ() {
-    int last = Integer.MIN_VALUE;
+    long last = Long.MIN_VALUE;
     for (int z = 0; z <= 15; z++) {
-      int encoded = TileCoord.ofXYZ(0, 0, z).encoded();
+      long encoded = TileCoord.ofXYZ(0, 0, z).encoded();
       if (encoded < last) {
         fail("encoded value for z" + (z - 1) + " (" + last + ") is not less than z" + z + " (" + encoded + ")");
       }
@@ -84,8 +84,8 @@ class TileCoordTest {
     "2,0,2,19",
     "3,0,2,20"
   })
-  void testTileCoordHilbert(int x, int y, int z, int i) {
-    int encoded = TileCoord.ofXYZ(x, y, z).hilbertEncoded();
+  void testTileCoordHilbert(int x, int y, int z, long i) {
+    long encoded = TileCoord.ofXYZ(x, y, z).hilbertEncoded();
     assertEquals(i, encoded);
     TileCoord decoded = TileCoord.hilbertDecode(i);
     assertEquals(decoded.x(), x, "x");
